Correction de l'architecture MVVM pour le SaveInformation & FileInformation
