+++ Livrable1/ViewModel/AddBackupViewModel.cs
@@ -21,100 +21,7 @@ using System.Windows.Media;
 namespace Livrable1.ViewModel
 {
 
-    //---------------------Model---------------------//
-    public class SaveInformation
-    {
-        //---------Variable---------//
-        public string NameSave { get; set; } = "";
-        public string CheminSource { get; set; } = "";
-        public string CheminDestination { get; set; } = "";
-        public List<FileInformation> Files { get; set; } = new List<FileInformation>();
-        //---------Variable---------//
-
-
-        //---------Constructeur---------//
-        public SaveInformation(string nameSave, string cheminSource, string cheminDestination)
-        {
-            NameSave = nameSave;
-            CheminSource = cheminSource;
-            CheminDestination = cheminDestination;
-            LoadFiles();
-        }
-        //---------Constructeur---------//
-
-
-        //---------Méthodes pour valider les chemins---------//
-        public bool ValidatePaths()
-        {
-            // Vérifier si le chemin source existe
-            if (!Directory.Exists(CheminSource))
-            {
-                MessageBox.Show("Le chemin source n'existe pas.");
-                return false;
-            }
-
-            // Vérifier si le chemin de destination existe
-            if (!Directory.Exists(CheminDestination))
-            {
-                MessageBox.Show("Le chemin de destination n'existe pas.");
-                return false;
-            }
-
-            // Vérifier que le chemin source contient des fichiers
-            string[] files = Directory.GetFiles(CheminSource);
-            if (files == null || files.Length == 0)
-            {
-                MessageBox.Show("Le chemin source ne contient aucun fichier.");
-                return false;
-            }
-
-            return true;
-        }
-        //---------Méthodes pour valider les chemins---------//
-
-
-        //---------Méthodes pour charger les fichiers---------//
-        public void LoadFiles()
-        {
-            Files.Clear(); // Réinitialise la liste
-            if (Directory.Exists(CheminSource))
-            {
-                string[] filePaths = Directory.GetFiles(CheminSource);
-                foreach (string filePath in filePaths)
-                {
-                    Files.Add(new FileInformation(filePath));
-                }
-            }
-        }
-        //---------Méthodes pour charger les fichiers---------//
-
-
-        //---------Méthodes pour sélectionner les fichiers---------//
-        public void SetSelectedFiles(List<FileInformation> selectedFiles)
-        {
-            Files = selectedFiles;
-        }
-        //---------Méthodes pour sélectionner les fichiers---------//
-
-    }
-    //---------------------------//
-    public class FileInformation
-    {
-        //---------Variable---------//
-        public string FilePath { get; set; }
-        public string FileName => Path.GetFileName(FilePath);
-        public bool IsSelected { get; set; } // Pour suivre la sélection
-        //---------Variable---------//
-
-        //---------Constructeur---------//
-        public FileInformation(string filePath)
-        {
-            FilePath = filePath;
-        }
-        //---------Constructeur---------//
-
-    }
-    //---------------------Model---------------------//
+    
 
 
     //---------------------ViewModel---------------------//
